requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple
--trusted-host mirror.aliyun.com

adbutils==2.9.2
    # via -r requirements.in
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
decorator==5.2.1
    # via retry2
deprecation==2.1.0
    # via adbutils
idna==3.10
    # via requests
numpy==2.2.6
    # via opencv-python
opencv-python==4.11.0.86
    # via -r requirements.in
packaging==25.0
    # via deprecation
pillow==11.2.1
    # via adbutils
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via adbutils
retry2==0.9.5
    # via adbutils
urllib3==2.4.0
    # via requests
